Tugas 4

+++ todolist/views.py
@@ -69,21 +69,31 @@ def create_task(request):
         task_baru.date = str(datetime.date.today())
         task_baru.title = request.POST.get('title')
         task_baru.description = request.POST.get('description')
+
         task_baru.save()
         messages.success(request, 'Task baru telah berhasil ditambah!')
         return redirect('todolist:show_todolist')
+
     context = {}
     return render(request, 'create-task.html', context)
 
 def update_task(request, id):
     task = Task.objects.get(id=id)
-    task.is_finished ^= True
+    if task.user == request.user:
+        task.is_finished ^= True
+    else:
+        return redirect('todolist:show_todolist')
+    
     task.save()
     messages.success(request, 'Status task telah berhasil diubah!')
     return redirect('todolist:show_todolist')
 
 def delete_task(request, id):
     task = Task.objects.get(id=id)
-    task.delete()
+    if task.user == request.user:
+        task.delete()
+    else:
+        return redirect('todolist:show_todolist')
+
     messages.success(request, 'Task telah berhasil dihapus!')
     return redirect('todolist:show_todolist')
